move word token helper generators to generate
